Move qpd inside circuit_cutting
