+++ circuit_knitting_toolbox/circuit_cutting/qpd/instructions/qpd_gate.py
@@ -18,7 +18,7 @@ from abc import ABC, abstractmethod
 
 from qiskit.circuit import QuantumCircuit, Instruction, CircuitInstruction
 
-from circuit_knitting_toolbox.qpd import QPDBasis
+from circuit_knitting_toolbox.circuit_cutting.qpd import QPDBasis
 
 
 class BaseQPDGate(Instruction, ABC):
